+++ src/common/index.ts
@@ -1,4 +1,3 @@
 export * from './common.constants.js';
 export * from './common.enums.js';
 export * from './common.interfaces.js';
-export * from './common.types.js';